requirements.txt
flask>=3.0.0
tensorflow>=2.20.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
werkzeug>=3.0.0